convert jpg images to webp and move to images dir
