+++ src/const/public/hippotherapy-page.ts
@@ -1,6 +1,6 @@
-import manAndHorse from '@/assets/images/public/about-us-page/man-horse.jpg';
-import manGrayHorse from '@/assets/images/public/hippotherapy/man_gray_horse.jpg';
-import girlAndHorse from '@/assets/images/public/about-us-page/girl-horse.jpg';
-import manLeadsHorse from '@/assets/images/public/hippotherapy/man_leads_horse.jpg';
+import manAndHorse from '@/assets/images/man-horse.webp';
+import manGrayHorse from '@/assets/images/man-gray-horse.webp';
+import girlAndHorse from '@/assets/images/collected.webp';
+import manLeadsHorse from '@/assets/images/man-leads-horse.webp';
 
 export const HIPPOTHERAPY_SWIPED_IMAGES = [manAndHorse, manGrayHorse, girlAndHorse, manLeadsHorse];
