+++ src/utils/mock-data/public/hippotherapy.ts
@@ -1,8 +1,8 @@
 import introImg from '@/assets/images/man-touching-white-horse-side.webp';
-import advantageImage0 from '@/assets/images/man-in-stall-stroking-horse.webp';
-import advantageImage1 from '@/assets/images/man-in-stall-stroking-horse.webp';
-import advantageImage2 from '@/assets/images/man-in-stall-stroking-horse.webp';
-import advantageImage3 from '@/assets/images/man-in-stall-stroking-horse.webp';
+import advantageImage0 from '@/assets/images/man-horse.webp';
+import advantageImage1 from '@/assets/images/man-gray-horse.webp';
+import advantageImage2 from '@/assets/images/collected.webp';
+import advantageImage3 from '@/assets/images/man-leads-horse.webp';
 import quoteImg from '@/assets/images/man-holding-horse-cheek-close.webp';
 import anotherQuoteImg from '@/assets/images/man-facing-horse-forehead.webp';
 import hippoventionImg from '@/assets/images/people-on-horses.webp';
